refactor: extract manager.ts into smaller modules
- Extract utilities to telemetry-utils.ts and span-utils.ts
- Extract module loading to module-loader.ts
- Extract exporters to exporters/ directory
- Extract metrics management to metrics-manager.ts
- Extract SDK initialization to sdk-initializer.ts
- Reduce manager.ts from 1273 to 360 lines (71% reduction)
- Improve maintainability and testability
