+++ packages/telemetry/src/otel/exporters/index.ts
@@ -1,0 +1,3 @@
+// packages/telemetry/src/otel/exporters/index.ts
+export { SafeOTLPMetricExporter } from './safe-otlp-metric-exporter';
+export { SafeOTLPExporter } from './safe-otlp-exporter';
